Share fix_province_name and fix_municipality_name between commands
