+++ pombola/south_africa/management/commands/south_africa_update_constituency_offices.py
@@ -31,7 +31,7 @@ from pombola.core.models import (OrganisationKind, Organisation, Place, PlaceKin
                          InformationSource)
 
 from mapit.models import Generation, Area, Code
-
+from ..helpers import fix_municipality_name
 
 class LocationNotFound(Exception):
     pass
@@ -56,19 +56,6 @@ test = 'yes'
 locationsnotfound = []
 personnotfound = []
 
-
-def fix_province_name(province_name):
-    if province_name == 'Kwa-Zulu Natal':
-        return 'KwaZulu-Natal'
-    else:
-        return province_name
-
-
-def fix_municipality_name(municipality_name):
-    if municipality_name == 'Merafong':
-        return 'Merafong City'
-    else:
-        return municipality_name
 
 
 def geocode(address_string, geocode_cache=None):
