+++ pombola/south_africa/management/helpers.py
@@ -1,0 +1,12 @@
+def fix_province_name(province_name):
+    if province_name == 'Kwa-Zulu Natal':
+        return 'KwaZulu-Natal'
+    else:
+        return province_name
+
+
+def fix_municipality_name(municipality_name):
+    if municipality_name == 'Merafong':
+        return 'Merafong City'
+    else:
+        return municipality_name
